Add files via upload
This bash script is designed for Ubuntu (tested to work in 14.04.1) which downloads packages required for integrating Ubuntu into a Windows Domain.  This script has also been tested and works on Linux Mint 17.3.

This script MUST be run as the root/sudo as the commands to acquire a Kerberos ticket and to join the domain will fail otherwise.